+++ packages/dataviews/src/utils/filter-sort-and-paginate.ts
@@ -389,7 +389,10 @@ export default function filterSortAndPaginate< Item >(
 	// Handle sorting.
 	const sortByField = view.sort?.field
 		? _fields.find( ( field ) => {
-				return field.id === view.sort?.field;
+				return (
+					field.enableSorting !== false &&
+					field.id === view.sort?.field
+				);
 		  } )
 		: null;
 	const groupByField = view.groupBy?.field